requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-celery-results==1.1.2
django-redis==4.10.0
django-tinymce==2.8.0
dnspython==1.16.0
eventlet==0.25.0
fdfs-client-py==1.2.6
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
kombu==4.6.3
monotonic==1.5
mutagen==1.42.0
mysqlclient==1.4.2.post1
Pillow==6.1.0
PyMySQL==0.9.3
pytz==2019.1
redis==3.2.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
